Change model loading to use bfloat16 dtype

+++ src/Experiments/Logit/PCA.py
@@ -21,7 +21,7 @@ dataset = build_dataset(n=100, low=1000, high=9999)
 #model_name = "gemma-2-9b-it"
 model_name = "llama3-8b-it"
 
-model = load_model(model_name)
+model = load_model(model_name, torch_dtype="bfloat16")
 model.set_use_attn_result(True)
 device = get_device()
